Standard Developer Experience
I added a new document to the Developer Experience section describing
the use of a go script or program to implement common tasks that
software developers perform. The goal of a go script is to minimize the
amount of information a new developer needs to know to be productive in
building and running the product.