Standard Developer Experience (#30)
I added a new document to the Developer Experience section describing
the use of a go script or program to implement common tasks that
software developers perform. The goal of a go script is to minimize the
amount of information a new developer needs to know to be productive in
building and running the product.

I updated the Developer Experience section of the Engineering
Fundamentals Checklist to replace the setup, build, test, and run tasks
with references to the proposed `go` commands that were added in the
Standard Developer Experience document in the Developer Experience
section.